Generate HTML tables instead of rst
They are far more flexible, and this amount of data
needs a lot of careful formatting...
Instead of stdout the output goes to separate files,
so they can be included one by one in the docs
(this is required because Sphinx needs to know about
headers to create the page TOC).
git is invoked on catapult-rest to obtain its version.
A CSV report from symbol-bootstrap is required to
obtain the TESTNET default values.

Only the properties command has been tested. The others
might be broken. Coming soon.